requirements.txt
black==23.1.0
certifi==2022.12.7
click==8.1.3
elabapi-python==0.1.7
et-xmlfile==1.1.0
mypy-extensions==1.0.0
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
